Reorganize the modules
