+++ src/pycture/__init__.py
@@ -4,15 +4,15 @@ from PyQt5.QtWidgets import QMainWindow, QWidget, QGridLayout
 from PyQt5.QtGui import QPixmap
 from PyQt5.QtCore import QEvent
 
-from classes.controls_bar import ControlsBar
-from classes.editor import Editor
-from classes.events import NewEditorEvent, DeleteEditorEvent, ChangeActiveEditorEvent
+from .menu_bar import MenuBar
+from .editor import Editor
+from .events import NewEditorEvent, DeleteEditorEvent, ChangeActiveEditorEvent
 
 class MainWindow(QMainWindow):
     def __init__(self):
         super().__init__()
         self.setWindowTitle("Pycture")
-        self.setMenuBar(ControlsBar(self))
+        self.setMenuBar(MenuBar(self))
         placeholder_widget = QWidget()
         placeholder_widget.setLayout(QGridLayout())
         self.setCentralWidget(placeholder_widget)
@@ -24,7 +24,7 @@ class MainWindow(QMainWindow):
             event.command.execute(self)
         elif type(event) == DeleteEditorEvent:
             self.editors.pop(event.editor_name)
-        elif type(event) == ChangeFocusedEditorEvent:
+        elif type(event) == ChangeActiveEditorEvent:
             self.activeEditor = event.editor_name
         else:
             event.ignore()
